requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
pg8000
Flask-Migrate==4.0.4
python-dotenv==1.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Werkzeug==2.3.6
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
requests
google-auth

# MongoDB
pymongo
twilio
pymongo 
